+++ lib/pages/dashboards/dashboards_list_page.dart
@@ -1,99 +1,31 @@
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:lotti/classes/entity_definitions.dart';
-import 'package:lotti/database/database.dart';
-import 'package:lotti/get_it.dart';
-import 'package:lotti/utils/sort.dart';
-import 'package:lotti/widgets/app_bar/dashboards_app_bar.dart';
-import 'package:lotti/widgets/charts/empty_dashboards_widget.dart';
+import 'package:lotti/blocs/dashboards/dashboards_page_cubit.dart';
+import 'package:lotti/widgets/dashboards/dashboards_app_bar.dart';
+import 'package:lotti/widgets/dashboards/dashboards_list.dart';
 import 'package:lotti/widgets/habits/habit_page_app_bar.dart';
-import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
-import 'package:lotti/widgets/settings/settings_card.dart';
 
-class DashboardsListPage extends StatefulWidget {
+class DashboardsListPage extends StatelessWidget {
   const DashboardsListPage({super.key});
-
-  @override
-  State<DashboardsListPage> createState() => _DashboardsListPageState();
-}
-
-class _DashboardsListPageState extends State<DashboardsListPage> {
-  final JournalDb _db = getIt<JournalDb>();
-  late final Stream<List<DashboardDefinition>> stream = _db.watchDashboards();
-  String match = '';
-
-  @override
-  void initState() {
-    super.initState();
-  }
 
   @override
   Widget build(BuildContext context) {
     final localizations = AppLocalizations.of(context)!;
 
-    return StreamBuilder<List<DashboardDefinition>>(
-      stream: stream,
-      builder: (
-        BuildContext context,
-        AsyncSnapshot<List<DashboardDefinition>> snapshot,
-      ) {
-        if (snapshot.data == null) {
-          return const LoadingDashboards();
-        }
-
-        final dashboards = filteredSortedDashboards(
-          snapshot.data ?? [],
-          match: match,
-        );
-
-        return Scaffold(
-          body: SafeArea(
-            child: CustomScrollView(
-              slivers: <Widget>[
-                SliverTitleBar(localizations.navTabTitleInsights),
-                const DashboardsSliverAppBar(),
-                SliverToBoxAdapter(
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 5),
-                    child: Column(
-                      children: [
-                        ...dashboards.mapIndexed(
-                          (index, dashboard) => DashboardCard(
-                            dashboard: dashboard,
-                            index: index,
-                          ),
-                        )
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
+    return BlocProvider<DashboardsPageCubit>(
+      create: (BuildContext context) => DashboardsPageCubit(),
+      child: Scaffold(
+        body: SafeArea(
+          child: CustomScrollView(
+            slivers: <Widget>[
+              SliverTitleBar(localizations.navTabTitleInsights),
+              const DashboardsSliverAppBar(),
+              const DashboardsList(),
+            ],
           ),
-        );
-      },
-    );
-  }
-}
-
-class DashboardCard extends StatelessWidget {
-  const DashboardCard({
-    required this.dashboard,
-    required this.index,
-    super.key,
-  });
-
-  final DashboardDefinition dashboard;
-  final int index;
-
-  @override
-  Widget build(BuildContext context) {
-    return SettingsNavCard(
-      path: '/dashboards/${dashboard.id}',
-      title: dashboard.name,
-      contentPadding: contentPaddingWithLeading,
-      leading: CategoryColorIcon(dashboard.categoryId),
+        ),
+      ),
     );
   }
 }