+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Upgraded dependencies
+- Refactoring: BloC for state management on dashboards page
 
 ## [0.8.343] - 2023-05-04
 ### Fixed: